Fix a bug where we assumed standasrd streams were lvalues
In filter.c, after a fork, we try to reinit the standard stream
pointers stdin, stdout, and stderr -- which is a bug, since the macros
stdin, stdout, and sterr may not be modifiable lvalues (C99 says they
nmeed not be modifuable). This is also pointless, since POSIX
guarantees that the child inherits the streams from the parent. There
are no test failures from thie fix.

git-archimport-id: srivasta@debian.org--2003-primary/flex--stream-ptr-fix--2.5--patch-1