allow users access to the string "str" in #defines.
changed the implementation of the yy_scan_string function to use
yy_str, as stated in the declaration, so as to allow users access to
the string "str" in #defines. I am not completely happy with this fix,
since there are other variables that may also pollute the #define name
space, but going through and modifying them all would be a significant
departure from upstream code; and I'm going to wait until someone
complains.

git-archimport-id: srivasta@debian.org--2003-primary/flex--str-fix--2.5--patch-1