flex generates C program with invalid macro call
The problem is that the code generates a macro definition that looks
like this:

#define gram_wrap(n) 1

but it always invokes the macro with zero arguments.  This does not
conform to C89, and some picky compilers reject it.  For example,
please see this bug report:

http://mail.gnu.org/archive/html/bug-bison/2004-02/msg00026.html

which mentions this error message:

scan-gram.c:2256: only 0 args to macro 'gram_wrap' (1 expected)

git-archimport-id: srivasta@debian.org--2003-primary/flex--debian--2.5--patch-2