Add a static library compiled with position independent code
Starting with version 2.5.31-18 of flex we have started providing
a static library compiled with position independent code, namely,
libfl_pic.a.  The problem is with packages that contain share
libraries with a flex scanner compiled in. Since flex generates code
that is not self contained, and the missing symbols live in
libfl.a. However, since linking a shared library with a object
containing non position independent code stopped working with gcc 4.1
(apparently, it was sheer luck that it worked at all). So now we also
provide libfla_pic.a for shared library packages to link with.

An alternative would have been to provide a full fledged
shared library, but that would have been a major break with upstream
and other distributions that also have flex, and would entail
creating new library and devel packages for flex, which is more work
that seems justified (considering that we would have to maintain the
fork, and keep track of any ABI changes, and all that).

git-archimport-id: srivasta@debian.org--etch/flex--autotools-refresh--2.5.33--patch-2