Narrow buffer size changes to just IA64
Increasing the buffer size on other architectures cause problems in
corner cases not caught by the test suite. So perhaps it is not better
to increase the buffer size unilatereally for all architectures, but
to just increase it for the architecture it is required on, namely,
IA64, and leave it unchanged on other architectures. This solution
should fix the new grave bugs, while not reopening the original but
that cause this change to be done in the first palce.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>