[fix-off-by-one-error]: generatred line numbers are off by one
The code that is generated by flex has an off-by-one error in the
generated #line pragmas that occur after the second %% line in the
source file (that is, the one that is used after the token patterns). As
a result, gdb will think we're, e.g., on line 46 when in reality we're
on line 45.

Marcus Rausch  supplied the patch in this commit

Signed-off-by: Manoj Srivastava <srivasta@golden-gryphon.com>