Really fix C++ comments in C parsers bug.
So, when we applied the upstream patch, it only fixed the skeleton
file, flex.skl, and not the generated files skel.c and scan.c;. Added
a feature branch that fixes the generated files, the need for this
will go away with the next upstream release. I hope. For what it is
worth, the diff in upstream seems to be the same one that was in the
NMU; and I have no explanations for why the -6 would behave diferently
from the NMU. (Closes: #813256).

Signed-off-by: Manoj Srivastava <srivasta@debian.org>